requirements.txt
streamlit
pandas
langchain
faiss-cpu
google-generativeai
streamlit
pandas
scikit-learn
sentence-transformers
google-generativeai
sentence-transformers
transformers
torch

altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
faiss-cpu==1.10.0
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.33.0
networkx==3.4.2
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.51.0
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
pyee==12.1.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.30.0
sentence-transformers==4.0.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
streamlit==1.44.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
requests

